Add files via upload
The 2030 Agenda for Sustainable Development was adopted by governments in 2015. Central to that Agenda is making progress on 17 Sustainable Development Goals (SDGs). Implemented at a time when globalisation is being buffeted by populism, a pandemic and its fallout, intensifying geopolitical rivalry, and attendant supply chain reconfiguration, the contribution of government commercial policies to attaining these Goals is in the spotlight. Analysing trends in global trade allows for a novel angle on the progress on these goals. From: https://www.globaltradealert.org/reports/gta-30-report